New Dynamatic to legacy Dynamatic bridge
This commit introduces a script that bridges new Dynamatic with legacy
Dynamatic. The script takes as input the handshake-level IR produced by
the new Dynamatic and preprocesses it before obtaining a DOT file
compatible with legacy Dynamatic. That DOT file is then converted to a
VHDL design by legacy Dynamatic.