[wlf2mp4] Prototype wlf2mp4 debugging sricpt
This commit introduces a Python script that helps debugging deadlocks
in dataflow circuits by creating a video out of a DOT graph that was
fed to Dynamatic's simulation flow and of a Modelsim output waveform
file (*.wlf) obtained from the simulation of the DOT with specific
inputs. The script outputs a series of DOTs, corresponding images, and
video (obtained from the concatenation of all images) showcasing the
valid/ready state of each channel in the dataflow circuit, allowing to
tack the flow of data over the simulation.