[Runner] Smart buffer placement and improvements
This commit mainly adds a smart buffer placement option in the script,
allowing to toggle between trivial and smart buffer placement strategies
in the legacy and bridge flows. It also adds an option to avoid
recompilation of benchmarks, instead using the cached DOTs of each
specified benchmark when present.

This also makes various (mostly cosmetic) small improvements to the
script.