[Runner] Better output handling
This commit improves the way the runner script handles its outputs,
making it possible to define a custom output path (as a subdirectory
of each benchmark) for a specific run thanks to a command line argument
and generally making the creation (and destruction) of output
directories more steamlined.

It also contains cosmetic changes and improves documentation throughout
the script.