[Viz] Visualizer for utilization/timing reports
This commit introduces the first version of a Python script that
creates charts to visualize data from the utilization and timing
reports created by Vivado's synthesization of our benchmarks.

In the future, the script will also be able to parse and visualize
simulation reports from Modelsim, as well as elastic circuits extracted
from DOTs.