[Build] Massive improvement to legacy setup
This commit largely expands and improves the setup_legacy_dynamatic.sh
script so that it can clone and configure legacy Dynamatic automatically
for use with (at this point) the runner.sh script. All necessary
binaries are built by the script and placed in the right locations so
that the paths indicated in the .env file correctly point to them. The
script should also be idempotent. This should make the process of using
legacy Dynamatic on this repository painless*.

*see terms and conditions