[Runner] Use benchmarks from Dynamatic repository
This commit updates the runner script to reflect recent changes in
Dynamatic, importantly the addition of integration tests as part of the
repository. As a consequence--and to maintain a single "source of truth"
for our benchmarks--the concept of "testsuite" is removed from the
runner script, which now always consider that it should go look in the
Dynamatic repository's "integration-test" directory for benchmarks.

This makes the local benchmarks maintained by this repository useless,
so they are removed as part of this commit.